feat: add filters

+++ app/domain/paging.go
@@ -1,7 +1,7 @@
 package domain
 
 type Paging struct {
-	Limit  int64 `json:"limit"`
-	Offset int64 `json:"offset"`
-	Total  int64 `json:"total"`
+	Limit  int `json:"limit"`
+	Offset int `json:"offset"`
+	Total  int `json:"total"`
 }
